@@ -1,24 +1,23 @@
-﻿namespace GLFW
+﻿namespace GLFW;
+
+/// <summary>
+///     Strongly-typed values for commonly used constants.
+///     <para>You are free to use the integer value instead of these constants.</para>
+/// </summary>
+public enum Constants
 {
-	/// <summary>
-	///     Strongly-typed values for commonly used constants.
-	///     <para>You are free to use the integer value instead of these constants.</para>
-	/// </summary>
-	public enum Constants
-    {
-        /// <summary>
-        ///     No preference or don't care, use default value.
-        /// </summary>
-        Default = -1,
+    /// <summary>
+    ///     No preference or don't care, use default value.
+    /// </summary>
+    Default = -1,
 
-        /// <summary>
-        ///     Boolean false
-        /// </summary>
-        False = 0,
+    /// <summary>
+    ///     Boolean false
+    /// </summary>
+    False = 0,
 
-        /// <summary>
-        ///     Boolean true
-        /// </summary>
-        True = 1
-    }
+    /// <summary>
+    ///     Boolean true
+    /// </summary>
+    True = 1
 }
