@@ -1,17 +1,16 @@
 #pragma warning disable 1591
 
-namespace GLFW
+namespace GLFW;
+
+/// <summary>
+///     Represents a gamepad axis.
+/// </summary>
+public enum GamePadAxis
 {
-    /// <summary>
-    ///     Represents a gamepad axis.
-    /// </summary>
-    public enum GamePadAxis
-    {
-        LeftX = 0,
-        LeftY = 1,
-        RightX = 2,
-        RightY = 3,
-        LeftTrigger = 4,
-        RightTrigger = 5
-    }
+    LeftX = 0,
+    LeftY = 1,
+    RightX = 2,
+    RightY = 3,
+    LeftTrigger = 4,
+    RightTrigger = 5
 }
