@@ -1,88 +1,87 @@
-﻿namespace GLFW
+﻿namespace GLFW;
+
+/// <summary>
+///     Strongly-typed values describing possible joysticks.
+/// </summary>
+public enum Joystick
 {
     /// <summary>
-    ///     Strongly-typed values describing possible joysticks.
+    ///     Joystick 1
     /// </summary>
-    public enum Joystick
-    {
-        /// <summary>
-        ///     Joystick 1
-        /// </summary>
-        Joystick1 = 0,
+    Joystick1 = 0,
 
-        /// <summary>
-        ///     Joystick 2
-        /// </summary>
-        Joystick2 = 1,
+    /// <summary>
+    ///     Joystick 2
+    /// </summary>
+    Joystick2 = 1,
 
-        /// <summary>
-        ///     Joystick 3
-        /// </summary>
-        Joystick3 = 2,
+    /// <summary>
+    ///     Joystick 3
+    /// </summary>
+    Joystick3 = 2,
 
-        /// <summary>
-        ///     Joystick 4
-        /// </summary>
-        Joystick4 = 3,
+    /// <summary>
+    ///     Joystick 4
+    /// </summary>
+    Joystick4 = 3,
 
-        /// <summary>
-        ///     Joystick 5
-        /// </summary>
-        Joystick5 = 4,
+    /// <summary>
+    ///     Joystick 5
+    /// </summary>
+    Joystick5 = 4,
 
-        /// <summary>
-        ///     Joystick 6
-        /// </summary>
-        Joystick6 = 5,
+    /// <summary>
+    ///     Joystick 6
+    /// </summary>
+    Joystick6 = 5,
 
-        /// <summary>
-        ///     Joystick 7
-        /// </summary>
-        Joystick7 = 6,
+    /// <summary>
+    ///     Joystick 7
+    /// </summary>
+    Joystick7 = 6,
 
-        /// <summary>
-        ///     Joystick 8
-        /// </summary>
-        Joystick8 = 7,
+    /// <summary>
+    ///     Joystick 8
+    /// </summary>
+    Joystick8 = 7,
 
-        /// <summary>
-        ///     Joystick 9
-        /// </summary>
-        Joystick9 = 8,
+    /// <summary>
+    ///     Joystick 9
+    /// </summary>
+    Joystick9 = 8,
 
-        /// <summary>
-        ///     Joystick 10
-        /// </summary>
-        Joystick10 = 9,
+    /// <summary>
+    ///     Joystick 10
+    /// </summary>
+    Joystick10 = 9,
 
-        /// <summary>
-        ///     Joystick 11
-        /// </summary>
-        Joystick11 = 10,
+    /// <summary>
+    ///     Joystick 11
+    /// </summary>
+    Joystick11 = 10,
 
-        /// <summary>
-        ///     Joystick 12
-        /// </summary>
-        Joystick12 = 11,
+    /// <summary>
+    ///     Joystick 12
+    /// </summary>
+    Joystick12 = 11,
 
-        /// <summary>
-        ///     Joystick 13
-        /// </summary>
-        Joystick13 = 12,
+    /// <summary>
+    ///     Joystick 13
+    /// </summary>
+    Joystick13 = 12,
 
-        /// <summary>
-        ///     Joystick 14
-        /// </summary>
-        Joystick14 = 13,
+    /// <summary>
+    ///     Joystick 14
+    /// </summary>
+    Joystick14 = 13,
 
-        /// <summary>
-        ///     Joystick 15
-        /// </summary>
-        Joystick15 = 14,
+    /// <summary>
+    ///     Joystick 15
+    /// </summary>
+    Joystick15 = 14,
 
-        /// <summary>
-        ///     Joystick 16
-        /// </summary>
-        Joystick16 = 15
-    }
+    /// <summary>
+    ///     Joystick 16
+    /// </summary>
+    Joystick16 = 15
 }
