@@ -1,10 +1,10 @@
-﻿namespace GLFW
+﻿namespace GLFW;
+
+/// <summary>
+///     Strongly-typed enumeration describing mouse buttons.
+/// </summary>
+public enum MouseButton
 {
-    /// <summary>
-    ///     Strongly-typed enumeration describing mouse buttons.
-    /// </summary>
-    public enum MouseButton
-    {
 	    /// <summary>
 	    ///     Mouse button 1.
 	    ///     <para>Same as <see cref="Left" />.</para>
@@ -23,47 +23,46 @@
 	    /// </summary>
 	    Button3 = 2,
 
-        /// <summary>
-        ///     Mouse button 4.
-        /// </summary>
-        Button4 = 3,
+    /// <summary>
+    ///     Mouse button 4.
+    /// </summary>
+    Button4 = 3,
 
-        /// <summary>
-        ///     Mouse button 4.
-        /// </summary>
-        Button5 = 4,
+    /// <summary>
+    ///     Mouse button 4.
+    /// </summary>
+    Button5 = 4,
 
-        /// <summary>
-        ///     Mouse button 5.
-        /// </summary>
-        Button6 = 5,
+    /// <summary>
+    ///     Mouse button 5.
+    /// </summary>
+    Button6 = 5,
 
-        /// <summary>
-        ///     Mouse button 6.
-        /// </summary>
-        Button7 = 6,
+    /// <summary>
+    ///     Mouse button 6.
+    /// </summary>
+    Button7 = 6,
 
-        /// <summary>
-        ///     Mouse button 7.
-        /// </summary>
-        Button8 = 7,
+    /// <summary>
+    ///     Mouse button 7.
+    /// </summary>
+    Button8 = 7,
 
-        /// <summary>
-        ///     The left mouse button.
-        ///     <para>Same as <see cref="Button1" />.</para>
-        /// </summary>
-        Left = Button1,
+    /// <summary>
+    ///     The left mouse button.
+    ///     <para>Same as <see cref="Button1" />.</para>
+    /// </summary>
+    Left = Button1,
 
-        /// <summary>
-        ///     The right mouse button.
-        ///     <para>Same as <see cref="Button2" />.</para>
-        /// </summary>
-        Right = Button2,
+    /// <summary>
+    ///     The right mouse button.
+    ///     <para>Same as <see cref="Button2" />.</para>
+    /// </summary>
+    Right = Button2,
 
-        /// <summary>
-        ///     The middle mouse button.
-        ///     <para>Same as <see cref="Button3" />.</para>
-        /// </summary>
-        Middle = Button3
-    }
+    /// <summary>
+    ///     The middle mouse button.
+    ///     <para>Same as <see cref="Button3" />.</para>
+    /// </summary>
+    Middle = Button3
 }
