@@ -1,51 +1,42 @@
-﻿namespace GLFW
+﻿namespace GLFW;
+
+/// <summary>
+/// Base exception class for GLFW related errors.
+/// </summary>
+public class Exception : System.Exception
 {
     /// <summary>
-    /// Base exception class for GLFW related errors.
+    ///     Generic error messages if only an error code is supplied as an argument to the constructor.
     /// </summary>
-    public class Exception : System.Exception
+    /// <param name="code">The error code.</param>
+    /// <returns>Error message.</returns>
+    public static string GetErrorMessage(ErrorCode code)
     {
-        #region Methods
-
-        /// <summary>
-        ///     Generic error messages if only an error code is supplied as an argument to the constructor.
-        /// </summary>
-        /// <param name="code">The error code.</param>
-        /// <returns>Error message.</returns>
-        public static string GetErrorMessage(ErrorCode code)
+        return code switch
         {
-            switch (code)
-            {
-                case ErrorCode.NotInitialized:     return Strings.NotInitialized;
-                case ErrorCode.NoCurrentContext:   return Strings.NoCurrentContext;
-                case ErrorCode.InvalidEnum:        return Strings.InvalidEnum;
-                case ErrorCode.InvalidValue:       return Strings.InvalidValue;
-                case ErrorCode.OutOfMemory:        return Strings.OutOfMemory;
-                case ErrorCode.ApiUnavailable:     return Strings.ApiUnavailable;
-                case ErrorCode.VersionUnavailable: return Strings.VersionUnavailable;
-                case ErrorCode.PlatformError:      return Strings.PlatformError;
-                case ErrorCode.FormatUnavailable:  return Strings.FormatUnavailable;
-                case ErrorCode.NoWindowContext:    return Strings.NoWindowContext;
-                default:                           return Strings.UnknownError;
-            }
-        }
-
-        #endregion
-
-        #region Constructors
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="Exception" /> class.
-        /// </summary>
-        /// <param name="error">The error code to create a generic message from.</param>
-        public Exception(ErrorCode error) : base(GetErrorMessage(error)) { }
-
-        /// <summary>
-        ///     Initializes a new instance of the <see cref="Exception" /> class.
-        /// </summary>
-        /// <param name="message">The error message.</param>
-        public Exception(string message) : base(message) { }
-
-        #endregion
+            ErrorCode.NotInitialized => Strings.NotInitialized,
+            ErrorCode.NoCurrentContext => Strings.NoCurrentContext,
+            ErrorCode.InvalidEnum => Strings.InvalidEnum,
+            ErrorCode.InvalidValue => Strings.InvalidValue,
+            ErrorCode.OutOfMemory => Strings.OutOfMemory,
+            ErrorCode.ApiUnavailable => Strings.ApiUnavailable,
+            ErrorCode.VersionUnavailable => Strings.VersionUnavailable,
+            ErrorCode.PlatformError => Strings.PlatformError,
+            ErrorCode.FormatUnavailable => Strings.FormatUnavailable,
+            ErrorCode.NoWindowContext => Strings.NoWindowContext,
+            _ => Strings.UnknownError,
+        };
     }
+
+    /// <summary>
+    ///     Initializes a new instance of the <see cref="Exception" /> class.
+    /// </summary>
+    /// <param name="error">The error code to create a generic message from.</param>
+    public Exception(ErrorCode error) : base(GetErrorMessage(error)) { }
+
+    /// <summary>
+    ///     Initializes a new instance of the <see cref="Exception" /> class.
+    /// </summary>
+    /// <param name="message">The error message.</param>
+    public Exception(string message) : base(message) { }
 }
